refactor into chuck_globals
