+++ src/core/chuck.lex
@@ -36,7 +36,7 @@ IS          (u|U|l|L)*
 // file: chuck.yy.c
 // desc: chuck lexer
 //
-// author: Ge Wang (gewang.cs.princeton.edu) - generated by lex
+// author: Ge Wang (gewang@cs.princeton.edu) - generated by lex
 //         Perry R. Cook (prc@cs.princeton.edu)
 //
 // based on the ansi C grammar by Jeff Lee, maintained by Jutta Degener
